LangChainAzureSearch/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.15.0
azure-search-documents==11.4.0
azure-storage-blob==12.19.1
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.6.4
deepdiff==6.7.1
Deprecated==1.2.14
distro==1.9.0
effdet==0.4.1
emoji==2.10.1
filelock==3.13.1
filetype==1.2.0
flatbuffers==24.3.6
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.6
importlib_metadata==7.0.2
iopath==0.1.10
isodate==0.6.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.22
layoutparser==0.3.4
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
mdurl==0.1.2
mpmath==1.3.0
msal==1.27.0
msal-extensions==1.1.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==1.13.3
opencv-python==4.9.0.80
ordered-set==4.1.0
orjson==3.9.15
packaging==23.2
pandas==2.2.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pikepdf==8.13.0
pillow==10.2.0
pillow_heif==0.15.0
portalocker==2.8.2
protobuf==4.25.3
pyarrow==15.0.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
pypdf==4.1.0
pypdfium2==4.27.0
pyreadline3==3.4.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
scipy==1.12.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
streamlit==1.31.1
streamlit-chat==0.1.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.6.0
timm==0.9.16
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
transformers==4.38.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
unstructured==0.12.5
unstructured-client==0.21.1
unstructured-inference==0.7.23
unstructured.pytesseract==0.3.12
urllib3==2.2.1
validators==0.22.0
watchdog==4.0.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
